Refactor library into folder structure
