+++ custom_components/anycubic_cloud/anycubic_cloud_api/data_models/consumable.py
