+++ custom_components/anycubic_cloud/anycubic_cloud_api/data_models/printer.py
@@ -2,9 +2,31 @@ from __future__ import annotations
 
 from typing import TYPE_CHECKING, Any
 
-from .anycubic_data_model_consumable import AnycubicConsumableData
-from .anycubic_data_model_files import AnycubicFile
-from .anycubic_data_model_printer_properties import (
+from ..const.enums import (
+    AnycubicFunctionID,
+    AnycubicPrinterMaterialType,
+    AnycubicPrintStatus,
+)
+from ..exceptions.error_strings import (
+    ErrorsDataParsing,
+    ErrorsGeneral,
+    ErrorsMQTTUpdate,
+)
+from ..exceptions.exceptions import (
+    AnycubicAPIError,
+    AnycubicDataParsingError,
+    AnycubicMQTTUnhandledData,
+    AnycubicMQTTUnknownUpdate,
+)
+from ..helpers.helpers import (
+    get_part_from_mqtt_topic,
+    time_duration_string_to_delta,
+    timedelta_to_dhm_string,
+    timedelta_to_total_hours,
+)
+from .consumable import AnycubicConsumableData
+from .files import AnycubicFile
+from .printer_properties import (
     AnycubicMachineColorInfo,
     AnycubicMachineData,
     AnycubicMachineExternalShelves,
@@ -14,37 +36,15 @@ from .anycubic_data_model_printer_properties import (
     AnycubicMaterialColor,
     AnycubicMultiColorBox,
 )
-from .anycubic_data_model_printing_settings import AnycubicPrintingSettings
-from .anycubic_data_model_project import AnycubicProject
-from .anycubic_enums import (
-    AnycubicFunctionID,
-    AnycubicPrinterMaterialType,
-    AnycubicPrintStatus,
-)
-from .anycubic_error_strings import (
-    ErrorsDataParsing,
-    ErrorsGeneral,
-    ErrorsMQTTUpdate,
-)
-from .anycubic_exceptions import (
-    AnycubicAPIError,
-    AnycubicDataParsingError,
-    AnycubicMQTTUnhandledData,
-    AnycubicMQTTUnknownUpdate,
-)
-from .anycubic_helpers import (
-    get_part_from_mqtt_topic,
-    time_duration_string_to_delta,
-    timedelta_to_dhm_string,
-    timedelta_to_total_hours,
-)
+from .printing_settings import AnycubicPrintingSettings
+from .project import AnycubicProject
 
 if TYPE_CHECKING:
     from datetime import timedelta
 
-    from .anycubic_api import AnycubicAPI
-    from .anycubic_data_model_print_response import AnycubicPrintResponse
-    from .anycubic_data_model_printer_properties import AnycubicDryingStatus, AnycubicMaterialMapping
+    from ..anycubic_api import AnycubicAPI
+    from .print_response import AnycubicPrintResponse
+    from .printer_properties import AnycubicDryingStatus, AnycubicMaterialMapping
 
 
 class AnycubicPrinter:
