+++ custom_components/anycubic_cloud/anycubic_cloud_api/data_models/printer_properties.py
@@ -2,8 +2,8 @@ from __future__ import annotations
 
 from typing import Any
 
-from .anycubic_error_strings import ErrorsDataParsing
-from .anycubic_exceptions import AnycubicDataParsingError
+from ..exceptions.error_strings import ErrorsDataParsing
+from ..exceptions.exceptions import AnycubicDataParsingError
 
 
 class AnycubicMaterialMapping:
