+++ custom_components/anycubic_cloud/helpers.py
@@ -7,7 +7,7 @@ from typing import TYPE_CHECKING, Any
 
 from homeassistant.helpers.device_registry import CONNECTION_NETWORK_MAC, DeviceInfo
 
-from .anycubic_cloud_api.anycubic_enums import AnycubicPrinterMaterialType
+from .anycubic_cloud_api.const.enums import AnycubicPrinterMaterialType
 from .const import (
     CONF_DRYING_PRESET_DURATION_,
     CONF_DRYING_PRESET_TEMPERATURE_,
