+++ custom_components/anycubic_cloud/scripts/script_base.py
@@ -7,7 +7,7 @@ from typing import Any
 
 import aiohttp
 
-from ..anycubic_cloud_api.anycubic_api_mqtt import AnycubicMQTTAPI
+from ..anycubic_cloud_api.anycubic_api import AnycubicMQTTAPI
 
 LOGGER = logging.getLogger(__name__)
 LOGGER.setLevel(logging.DEBUG)
