+++ custom_components/anycubic_cloud/services.py
@@ -17,9 +17,9 @@ from homeassistant.core import HomeAssistant, ServiceCall
 from homeassistant.exceptions import HomeAssistantError, ServiceValidationError
 from homeassistant.helpers import config_validation as cv, selector
 
-from .anycubic_cloud_api.anycubic_data_model_print_response import AnycubicPrintResponse
-from .anycubic_cloud_api.anycubic_data_model_printer import AnycubicPrinter
-from .anycubic_cloud_api.anycubic_data_model_printer_properties import AnycubicMaterialColor
+from .anycubic_cloud_api.data_models.print_response import AnycubicPrintResponse
+from .anycubic_cloud_api.data_models.printer import AnycubicPrinter
+from .anycubic_cloud_api.data_models.printer_properties import AnycubicMaterialColor
 from .const import (
     AC_EVENT_PRINT_CLOUD_START,
     ATTR_ANYCUBIC_EVENT,
